+++ common/src/commonIntegrationTest/kotlin/com/powersync/sync/SyncIntegrationTest.kt
@@ -2,6 +2,7 @@ package com.powersync.sync
 
 import app.cash.turbine.turbineScope
 import co.touchlab.kermit.ExperimentalKermitApi
+import co.touchlab.kermit.Severity
 import com.powersync.ExperimentalPowerSyncAPI
 import com.powersync.PowerSyncDatabase
 import com.powersync.PowerSyncException
@@ -36,6 +37,8 @@ import io.kotest.matchers.comparables.shouldBeLessThan
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
 import io.kotest.matchers.string.shouldContain
+import io.ktor.client.HttpClient
+import io.ktor.client.engine.mock.MockEngine
 import io.ktor.http.ContentType
 import io.ktor.utils.io.core.toByteArray
 import kotlinx.coroutines.CompletableDeferred
@@ -46,9 +49,11 @@ import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.test.testTimeSource
 import kotlinx.coroutines.withContext
+import kotlinx.serialization.json.JsonPrimitive
 import kotlinx.serialization.json.jsonObject
 import kotlinx.serialization.json.jsonPrimitive
 import kotlin.test.Test
+import kotlin.test.assertContains
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 import kotlin.test.assertNotNull
@@ -84,8 +89,10 @@ class SyncIntegrationTest : AbstractSyncTest() {
         databaseTest {
             database.connect(
                 connector,
-                options = getOptions(),
-                params = mapOf("foo" to JsonParam.String("bar")),
+                options =
+                    getOptions().copy(
+                        params = mapOf("foo" to JsonParam.String("bar")),
+                    ),
             )
             turbineScope(timeout = 10.0.seconds) {
                 val turbine = database.currentStatus.asFlow().testIn(this)
@@ -108,10 +115,12 @@ class SyncIntegrationTest : AbstractSyncTest() {
         databaseTest {
             database.connect(
                 connector,
-                options = getOptions(),
-                appMetadata =
-                    mapOf(
-                        "app_version" to "1.0.0",
+                options =
+                    getOptions().copy(
+                        appMetadata =
+                            mapOf(
+                                "app_version" to "1.0.0",
+                            ),
                     ),
             )
             turbineScope(timeout = 10.0.seconds) {
@@ -467,13 +476,13 @@ class SyncIntegrationTest : AbstractSyncTest() {
             turbineScope(timeout = 10.0.seconds) {
                 val turbine = database.currentStatus.asFlow().testIn(this)
 
-                database.connect(connector, 1000L, options = getOptions())
+                database.connect(connector, options = getOptions())
                 turbine.waitFor { it.connecting }
 
                 database.disconnect()
                 turbine.waitFor { !it.connecting }
 
-                database.connect(connector, 1000L, options = getOptions())
+                database.connect(connector, options = getOptions())
                 turbine.waitFor { it.connecting }
                 database.disconnect()
                 turbine.waitFor { !it.connecting }
@@ -488,13 +497,61 @@ class SyncIntegrationTest : AbstractSyncTest() {
             turbineScope(timeout = 10.0.seconds) {
                 val turbine = database.currentStatus.asFlow().testIn(this)
 
-                database.connect(connector, 1000L, retryDelayMs = 5000, options = getOptions())
+                database.connect(connector, getOptions())
                 turbine.waitFor { it.connecting }
 
-                database.connect(connector, 1000L, retryDelayMs = 5000, options = getOptions())
+                database.connect(connector, getOptions())
                 turbine.waitFor { it.connecting }
 
                 turbine.cancelAndIgnoreRemainingEvents()
+            }
+        }
+
+    @OptIn(ExperimentalPowerSyncAPI::class, ExperimentalKermitApi::class)
+    @Test
+    fun warnsWhenConnectorDoesNotUploadData() =
+        databaseTest {
+            val assertNoHttpEngine =
+                MockEngine { request ->
+                    error("Unexpected HTTP request: $request")
+                }
+
+            connector.uploadDataCallback = {}
+            database.execute("INSERT INTO users (id, name) VALUES (uuid(), ?)", listOf("local users"))
+
+            database.connect(
+                connector,
+                options =
+                    SyncOptions(
+                        retryDelay = 10.seconds,
+                        clientConfiguration =
+                            SyncClientConfiguration.ExistingClient(
+                                HttpClient(assertNoHttpEngine) {
+                                    configureSyncHttpClient()
+                                },
+                            ),
+                    ),
+            )
+
+            delay(2.seconds)
+            database.disconnect()
+
+            val logs = logWriter.logs.filter { it.message.contains("CRUD") }
+
+            with(logs[0]) {
+                assertContains(
+                    message,
+                    "Potentially previously uploaded CRUD entries are still present in the upload queue.",
+                )
+                assertEquals(Severity.Warn, severity)
+            }
+
+            with(logs[1]) {
+                assertEquals(
+                    message,
+                    "Error uploading crud: Delaying due to previously encountered CRUD item.",
+                )
+                assertEquals(Severity.Error, severity)
             }
         }
 
@@ -700,7 +757,7 @@ class SyncIntegrationTest : AbstractSyncTest() {
             turbineScope(timeout = 10.0.seconds) {
                 val turbine = database.currentStatus.asFlow().testIn(this)
 
-                database.connect(connector, 1000L, retryDelayMs = 5000, options = getOptions())
+                database.connect(connector, getOptions())
                 turbine.waitFor { it.connecting }
 
                 syncLines.send(SyncLine.KeepAlive(tokenExpiresIn = 4000))
@@ -740,7 +797,7 @@ class SyncIntegrationTest : AbstractSyncTest() {
             turbineScope(timeout = 10.0.seconds) {
                 val turbine = database.currentStatus.asFlow().testIn(this)
 
-                database.connect(connector, 1000L, retryDelayMs = 5000, options = getOptions())
+                database.connect(connector, getOptions())
                 turbine.waitFor { it.downloadError != null }
 
                 database.currentStatus.downloadError?.toString() shouldContain "Expected exception from fetchCredentials"
@@ -783,7 +840,7 @@ class SyncIntegrationTest : AbstractSyncTest() {
             turbineScope(timeout = 10.0.seconds) {
                 val turbine = database.currentStatus.asFlow().testIn(this)
 
-                database.connect(connector, 1000L, retryDelayMs = 5000, options = getOptions())
+                database.connect(connector, getOptions())
                 turbine.waitFor { it.connecting }
 
                 syncLines.send(SyncLine.KeepAlive(tokenExpiresIn = 4000))
@@ -1153,5 +1210,34 @@ class SyncIntegrationTest : AbstractSyncTest() {
 
             logWriter.logs.any { it.message.contains("previously uploaded CRUD entries are still present in the upload queue") } shouldBe
                 false
+        }
+
+    @Suppress("DEPRECATION")
+    @Test
+    fun `can use legacy parameters`() =
+        databaseTest {
+            database.connect(
+                connector,
+                options =
+                    getOptions().copy(
+                        appMetadata =
+                            mapOf(
+                                "option_1" to "1",
+                            ),
+                    ),
+                appMetadata = mapOf("option_2" to "2"),
+            )
+            turbineScope(timeout = 10.0.seconds) {
+                val turbine = database.currentStatus.asFlow().testIn(this)
+                turbine.waitFor { it.connected }
+                turbine.cancel()
+            }
+
+            requestedSyncStreams shouldHaveSingleElement {
+                val meta = it.jsonObject["app_metadata"]!!.jsonObject
+                meta["option_1"] shouldBe JsonPrimitive("1")
+                meta["option_2"] shouldBe JsonPrimitive("2")
+                true
+            }
         }
 }